# perf: open popup window and run hooks in a single execution