--- a/src/toggle_tests/open_nested_popup.stdout
+++ b/src/toggle_tests/open_nested_popup.stdout
@@ -4,22 +4,19 @@ before-open
 ;
 run
 #{@before_open}
-<<<TMUX:END(1)
-
->>>TMUX:BEGIN(2)
+;
 set
 default-shell
 /bin/sh
 ;
-popup
-tmux set default-shell '/usr/bin/fish'; export TMUX_POPUP_SERVER='popup_server2' SHELL='/usr/bin/fish'; exec tmux -S socket/path/popup_server2 new -As pane/path/p_open_nested_2 \; set @__popup_name p_open_nested_2 \; set @__popup_id_format pane/path/\{popup_name\} \; set @__popup_caller_path working/session/path \; set @__popup_caller_pane_path working/pane/path \; display on-init \; run \#\{@on_init\}  >/dev/null
-<<<TMUX:END(2)
-
->>>TMUX:BEGIN(3)
+display-popup
+tmux set default-shell '/usr/bin/fish' ;export TMUX_POPUP_SERVER='popup_server2' ;export SHELL='/usr/bin/fish' ;exec tmux -S socket/path/popup_server2 new -As pane/path/p_open_nested_2 \; set @__popup_name p_open_nested_2 \; set @__popup_id_format pane/path/\{popup_name\} \; set @__popup_caller_path working/session/path \; set @__popup_caller_pane_path working/pane/path \; display on-init \; run \#\{@on_init\}  >/dev/null
+;
 display
 after-close
 ;
 run
 #{@after_close}
-<<<TMUX:END(3)
+;
+<<<TMUX:END(1)