--- a/src/toggle_tests/open_with_toggle_key.stdout
+++ b/src/toggle_tests/open_with_toggle_key.stdout
@@ -4,18 +4,23 @@ before-open
 ;
 run
 #{@before_open}
-<<<TMUX:END(1)
-
->>>TMUX:BEGIN(2)
+;
 set
 default-shell
 /bin/sh
 ;
-popup
-tmux set default-shell '/usr/bin/fish'; export TMUX_POPUP_SERVER='popup_server2' SHELL='/usr/bin/fish'; exec tmux -S socket/path/popup_server2 new -As pane/path/p_toggle_key \; set @__popup_name p_toggle_key \; set @__popup_id_format pane/path/\{popup_name\} \; set @__popup_caller_path working/session/path \; set @__popup_caller_pane_path working/pane/path \; bind -T root M-p run \#\{@popup-toggle\}\ --name=p_toggle_key\ --toggle-key=-T\\\ root\\\ M-p\ --toggle-key=-n\\\ M-o\  \; bind -n M-o run \#\{@popup-toggle\}\ --name=p_toggle_key\ --toggle-key=-T\\\ root\\\ M-p\ --toggle-key=-n\\\ M-o\  \; display on-init \; run \#\{@on_init\}  >/dev/null
-<<<TMUX:END(2)
+display-popup
+tmux set default-shell '/usr/bin/fish' ;export TMUX_POPUP_SERVER='popup_server2' ;export SHELL='/usr/bin/fish' ;exec tmux -S socket/path/popup_server2 new -As pane/path/p_toggle_key \; set @__popup_name p_toggle_key \; set @__popup_id_format pane/path/\{popup_name\} \; set @__popup_caller_path working/session/path \; set @__popup_caller_pane_path working/pane/path \; bind -T root M-p run \#\{@popup-toggle\}\ --name=p_toggle_key\ --toggle-key=-T\\\ root\\\ M-p\ --toggle-key=-n\\\ M-o\  \; bind -n M-o run \#\{@popup-toggle\}\ --name=p_toggle_key\ --toggle-key=-T\\\ root\\\ M-p\ --toggle-key=-n\\\ M-o\  \; display on-init \; run \#\{@on_init\}  >/dev/null
+;
+display
+after-close
+;
+run
+#{@after_close}
+;
+<<<TMUX:END(1)
 
->>>TMUX:BEGIN(3)
+>>>TMUX:BEGIN(2)
 -N
 -S
 socket/path/popup_server2
@@ -28,13 +33,5 @@ unbind
 -n
 M-o
 ;
-<<<TMUX:END(3)
-
->>>TMUX:BEGIN(4)
-display
-after-close
-;
-run
-#{@after_close}
-<<<TMUX:END(4)
+<<<TMUX:END(2)